Safari-611.1.10/LayoutTests/http/tests/preconnect/link-rel-preconnect-https-expected.txt

CONSOLE MESSAGE: Successfuly preconnected to https://localhost:8443/
Tests that Link's rel=preconnect works as expected over HTTPS.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
